requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
psycopg2==2.7.3.1
pyserial==3.4
requests==2.18.4
SQLAlchemy==1.1.14
urllib3==1.22
